pull from Apache

+++ src/components/src/main/java/org/apache/jmeter/timers/BeanShellTimer.java
@@ -6,13 +6,13 @@
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  *
  */
 
@@ -62,7 +62,7 @@ public class BeanShellTimer extends BeanShellTestElement implements Cloneable, T
             }
         }
         try {
-            return Long.decode(ret).longValue();
+            return Long.decode(ret);
         } catch (NumberFormatException e){
             log.warn("Number format exception while decoding number: '{}'", ret);
             return 0;
